Enforce unique job keys (#50)
* Proper support for empty string values in rmap

Setting an empty string value then calling Get now
returns empty string and true.

Also remove the need for using a sentinel value ("*")
when doing a reset.

* Update dependencies

* Enforce unique job keys

This commit does a couple of things:
1. It adds a new `SetIfNotExists` methods to replicated maps that will set a value only if it isn't already in the map.
2. It ensures the pool node `DispatchJob` method returns an the new `ErrJobExists` error when attempting to queue a job with a key that already exists.

It was previously the responsibility of the client to ensure no two calls to DispatchJob used the same key, now the pool package enforces it.

* Encode rmap string content in publish

So that we can ensure decoding all the parts in Go.
This replaces the previous algo that was using : to separate key and values
which was not working if a key or a value also contained :.

* Protect rmap.SetAndWait against corner cases

Fix all tests